requirements.txt
glfw==2.10.0
numpy==2.3.3
pillow==11.3.0
pyglm==2.8.2
PyOpenGL==3.1.9